+++ chat-widget/src/components/livechatwidget/interfaces/IPersistentChatHistoryProps.ts
@@ -1,0 +1,8 @@
+import { IStyle } from "@fluentui/react";
+
+export interface IPersistentChatHistoryProps {
+        persistentChatHistoryEnabled?: boolean;
+        pageSize?: number;
+        dividerActivityStyle?: IStyle;
+        bannerStyle?: IStyle;
+}