+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/activityStatusMiddleware.tsx
@@ -1,6 +1,7 @@
 import { Constants, WebChatMiddlewareConstants } from "../../../../../common/Constants";
 
 import { DeliveredTimestamp } from "./timestamps/DeliveredTimestamp";
+import { HistoryMessageTimestamp } from "./timestamps/HistoryMessageTimestamp";
 import { NotDeliveredTimestamp } from "./timestamps/NotDeliveredTimestamp";
 import React from "react";
 import { SendStatus } from "../../enums/SendStatus";
@@ -25,6 +26,11 @@ export const activityStatusMiddleware = () => (next: any) => (args: any) => {
 
     const current_tags = tags, current_name = name, current_role = role, current_timestamp = timestamp;
     let sameTimestampGroupTemp: boolean = sameTimestampGroup;
+
+    if (tags?.includes(Constants.persistentChatHistoryMessageTag)) {
+        return <HistoryMessageTimestamp args={args} />;
+    }
+
     if (args[WebChatMiddlewareConstants.nextVisibleActivity]) {
         const {
             nextVisibleActivity: {
@@ -34,6 +40,7 @@ export const activityStatusMiddleware = () => (next: any) => (args: any) => {
                 timestamp
             }
         } = args;
+        
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         const next_name = name, next_timestamp = timestamp;
         const timestampCurrentInMillisecond = Date.parse(current_timestamp);
@@ -55,4 +62,9 @@ export const activityStatusMiddleware = () => (next: any) => (args: any) => {
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         ) as any;
     }
+};
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const createActivityStatusMiddleware = (locale = "en-us") => {
+    return activityStatusMiddleware;
 };