+++ CHANGE_LOG.md
@@ -6,8 +6,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Adding support to fetch history messages for persistent chat
+- Add use of `config.LcwFcbConfiguration?.lcwPersistentChatHistoryEnabled` to enable persistent chat history feature
+
+## [1.8.3] - 2025-10-07
+
 ### Fixed
 
+- IOS mobile doesnt display emoji when typing.
 - Fix 400% zoom issue on citation pane
 - Missing export for LiveChatWidgetMockType in index.ts
 - Fixed issue where link post chat survey not displayed